+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.3
+
+* Added a loading indicator for color picker
+* Added widget tests
+
 ## 1.0.1, 1.0.2
 
 * Minor README changes